Add a unittest based test driver.
To make testing ParseResolvConf easier I split it into two
functions. The first takes a filename and opens it, and then passes
it to a new function ParseResolvConfFromIterable which actually
does the parsing. This makes it possible to test the parsing code
without actually needing a valid file.

I added tests to check the A, AAAA, MX, SRV return values.